src/audiosummarizer/LLMS/geminillm.py
import os
from dotenv import load_dotenv
from langchain_google_genai import ChatGoogleGenerativeAI

# Load environment variables
load_dotenv()


class GeminiLLM:
    def __init__(self):
        self.model = "gemini-2.0-flash-exp"
        self.temperature = 0
        self.api_key = os.getenv("GEMINI_API_KEY")
        if not self.api_key:
            raise ValueError("GEMINI_API_KEY environment variable not set")
        
        self.llm = ChatGoogleGenerativeAI(
            model=self.model,
            temperature=self.temperature,
            google_api_key=self.api_key
        )

    def get_llm(self):
        return self.llm
    
